refactor: update ClientSessionLeaderThreadProxy usage across adapters and common types
