+++ packages/@livestore/common/src/leader-thread/recreate-db.ts
@@ -2,7 +2,8 @@ import { casesHandled } from '@livestore/utils'
 import type { HttpClient } from '@livestore/utils/effect'
 import { Effect, Queue } from '@livestore/utils/effect'
 
-import type { InvalidPullError, IsOfflineError, MigrationHooks, MigrationsReport, SqliteError } from '../index.js'
+import type { MigrationsReport } from '../defs.js'
+import type { InvalidPullError, IsOfflineError, MigrationHooks, SqliteError } from '../index.js'
 import { migrateDb, rematerializeFromEventlog, UnexpectedError } from '../index.js'
 import { configureConnection } from './connection.js'
 import { LeaderThreadCtx } from './types.js'
